requirements.txt
sentence-transformers
faiss-cpu
numpy
streamlit
requests
torchvision